+++ examples/ssr/hn2/src/views/UserView.vue
@@ -30,7 +30,7 @@ export default {
     },
     time () {
       const { value, unit } = timeAgo(this.user.created)
-      return `${value} ${this.$tc('time.units.' + unit, value)} ${this.$t('time.ago')}`
+      return `${value} ${this.$tp('time.units.' + unit, value)} ${this.$t('time.ago')}`
     }
   },